Allow multiple `=` characters in env values
This commit replaces the use of `split` in the environment variable
parsing with the first index of `=`, in order to properly handle
environment variables that contain `=` in their values.

Specifically, when parsing environment variables with multiple `=`
characters, the first one is used to separate the key from the value,
while the rest are kept as part of the value.
(For example, for the variable `FOO=bar=baz`, the key is `FOO`, and the
value is `bar=baz`).

Signed-off-by: Radu M <root@radu.sh>